(refactor) Move Inspection into its own file
There's no obvious place for this object to exist in the existing folder
structure. I chose 'Models' on the basis that this is an object
modelling the concept of a specific noun in our domain.

N.B. It's a data object, which is arguably a code smell:
https://sourcemaking.com/refactoring/smells/data-class
